Moving a lot of the functionality from the src/index.ts files, into its own modules
