+++ src/index.ts
@@ -1,162 +1,15 @@
 import { useMaskedErrors } from "@envelop/core";
 import { useImmediateIntrospection } from "@envelop/immediate-introspection";
 import { authZEnvelopPlugin } from "@graphql-authz/envelop-plugin";
-import { initContextCache } from "@pothos/core";
-import { decode, verify } from "@tsndr/cloudflare-worker-jwt";
-import { createGraphQLError, createYoga, maskError } from "graphql-yoga";
-import { Resend } from "resend";
+import { createYoga, maskError } from "graphql-yoga";
 
 import { Env } from "worker-configuration";
+import { logPossibleUserIdFromJWT } from "~/authn";
 import * as rules from "~/authz";
-import { ORM_TYPE, getDb } from "~/datasources/db";
-import { updateUserProfileInfo } from "~/datasources/queries/users";
+import { creageGraphqlContext } from "~/context";
 import { APP_ENV } from "~/env";
 import { logger } from "~/logging";
 import { schema } from "~/schema";
-
-import { Context } from "./builder";
-import { insertUsersSchema } from "./datasources/db/users";
-import { getMercadoPagoFetch } from "./datasources/mercadopago";
-import { getSanityClient } from "./datasources/sanity/client";
-import { getStripeClient } from "./datasources/stripe/client";
-
-// We get the token either from the Authorization header or from the "community-os-access-token" cookie
-const getAuthToken = (request: Request) => {
-  const cookieHeader = request.headers.get("Cookie");
-  const authHeader = request.headers.get("Authorization");
-  if (authHeader) {
-    const token = authHeader.split("Bearer ")[1];
-    return token;
-  }
-  if (cookieHeader) {
-    const cookies = cookieHeader.split(";").map((c) => c.trim());
-    const tokenCookie = cookies.find((c) =>
-      c.startsWith("community-os-access-token="),
-    );
-    if (tokenCookie) {
-      return tokenCookie.split("=")[1];
-    }
-  }
-  return null;
-};
-
-const decodeJWT = (JWT_TOKEN: string) => {
-  try {
-    const { payload } = decode(JWT_TOKEN) as {
-      payload: {
-        aud: string;
-        exp: number;
-        iat: number;
-        iss: string;
-        sub: string;
-        email: string;
-        phone: string;
-        app_metadata: { provider: string; providers: string[] };
-        user_metadata: {
-          avatar_url: string;
-          email: string;
-          email_verified: boolean;
-          full_name: string;
-          iss: string;
-          name: string;
-          phone_verified: boolean;
-          preferred_username: string;
-          provider_id: string;
-          picture: string;
-          sub: string;
-          user_name: string;
-        };
-        role: string;
-        aal: string;
-        amr: { method: "oauth"; timestamp: number }[];
-        session_id: string;
-        is_anonymous: boolean;
-      };
-    };
-    return payload;
-  } catch (e) {
-    logger.error("Could not parse token", e);
-    return null;
-  }
-};
-
-const unauthorizedError = (
-  message: string,
-  options?: Parameters<typeof createGraphQLError>[1],
-) => {
-  logger.error("Unauthorized Error:", message);
-  return createGraphQLError(message, {
-    extensions: {
-      code: "UNAUTHENTICATED",
-    },
-    ...options,
-  });
-};
-
-const getUser = async ({
-  request,
-  SUPABASE_JWT_DECODER,
-  DB,
-}: {
-  request: Request;
-  SUPABASE_JWT_DECODER: string;
-  DB: ORM_TYPE;
-}) => {
-  const JWT_TOKEN = getAuthToken(request);
-  if (!JWT_TOKEN) {
-    return null;
-  }
-  const payload = decodeJWT(JWT_TOKEN);
-  if (!payload) {
-    logger.error("Could not parse token");
-    return payload;
-  }
-  const verified = await verify(JWT_TOKEN, SUPABASE_JWT_DECODER);
-  if (!verified) {
-    throw unauthorizedError("Could not verify token");
-  }
-  const isExpired = payload.exp < Date.now() / 1000;
-  if (isExpired) {
-    throw unauthorizedError("Token expired");
-  }
-  const { avatar_url, name, user_name, email_verified, sub, picture } =
-    payload.user_metadata;
-  const profileInfo = insertUsersSchema.safeParse({
-    email: payload.email,
-    emailVerified: email_verified,
-    imageUrl: avatar_url ? avatar_url : picture ? picture : "",
-    externalId: sub,
-    name,
-    username: user_name,
-    publicMetadata: payload,
-  });
-  if (profileInfo.success === false) {
-    logger.error("Could not parse profile info", profileInfo.error);
-    throw new Error("Could not parse profile info", profileInfo.error);
-  }
-  logger.info("Updating profile Info for user ID:", sub);
-  return updateUserProfileInfo(DB, profileInfo.data);
-};
-
-const attachPossibleUserIdFromJWT = (request: Request) => {
-  const isOptions = request.method === "OPTIONS";
-  if (isOptions) {
-    return null;
-  }
-  const JWT_TOKEN = getAuthToken(request);
-  if (!JWT_TOKEN) {
-    logger.info("No token present");
-    return null;
-  }
-  try {
-    const { payload } = decode(JWT_TOKEN);
-    const userId = (payload as { id: string })?.id ?? "ANONYMOUS";
-    logger.info("User_ID", userId);
-  } catch (error) {
-    logger.error("Could not parse token", error);
-    return null;
-  }
-};
 
 export const yoga = createYoga<Env>({
   landingPage: APP_ENV !== "production",
@@ -208,104 +61,12 @@ export const yoga = createYoga<Env>({
     useImmediateIntrospection(),
     authZEnvelopPlugin({ rules }),
   ].filter(Boolean),
-  context: async ({
-    request,
-    NEON_URL,
-    PURCHASE_CALLBACK_URL,
-    MAIL_QUEUE,
-    GOOGLE_PHOTOS_IMPORT_QUEUE,
-    SANITY_PROJECT_ID,
-    SANITY_DATASET,
-    RESEND_API_KEY,
-    SANITY_API_VERSION,
-    SANITY_SECRET_TOKEN,
-    SUPABASE_JWT_DECODER,
-    STRIPE_KEY,
-    HYPERDRIVE,
-    MERCADOPAGO_KEY,
-  }) => {
-    if (!MAIL_QUEUE) {
-      throw new Error("Missing MAIL_QUEUE");
-    }
-    if (!GOOGLE_PHOTOS_IMPORT_QUEUE) {
-      throw new Error("Missing GOOGLE_PHOTOS_IMPORT_QUEUE");
-    }
-    if (!NEON_URL) {
-      throw new Error("Missing NEON_URL");
-    }
-    if (!SUPABASE_JWT_DECODER) {
-      throw new Error("Missing SUPABASE_JWT_DECODER");
-    }
-    if (!STRIPE_KEY) {
-      throw new Error("Missing STRIPE_KEY");
-    }
-    if (!MERCADOPAGO_KEY) {
-      throw new Error("Missing MERCADOPAGO_KEY");
-    }
-    if (!PURCHASE_CALLBACK_URL) {
-      throw new Error("Missing PURCHASE_CALLBACK_URL");
-    }
-    if (!RESEND_API_KEY) {
-      throw new Error("Missing RESEND_API_KEY");
-    }
-    if (
-      !SANITY_PROJECT_ID ||
-      !SANITY_DATASET ||
-      !SANITY_API_VERSION ||
-      !SANITY_SECRET_TOKEN
-    ) {
-      throw new Error("Missing Sanity Configuration");
-    }
-    const DB_URL =
-      HYPERDRIVE?.connectionString?.startsWith("postgresql://fake-user:fake") &&
-      APP_ENV === "development"
-        ? NEON_URL
-        : HYPERDRIVE.connectionString;
-
-    const GET_SANITY_CLIENT = () =>
-      getSanityClient({
-        projectId: SANITY_PROJECT_ID,
-        dataset: SANITY_DATASET,
-        apiVersion: SANITY_API_VERSION,
-        token: SANITY_SECRET_TOKEN,
-        useCdn: true,
-      });
-
-    const GET_STRIPE_CLIENT = () => getStripeClient(STRIPE_KEY);
-    const GET_MERCADOPAGO_CLIENT = getMercadoPagoFetch(MERCADOPAGO_KEY);
-    const DB = await getDb({
-      neonUrl: DB_URL,
-    });
-    const RESEND = new Resend(RESEND_API_KEY);
-    logger.info("Getting user");
-    const USER = await getUser({
-      request,
-      SUPABASE_JWT_DECODER,
-      DB,
-    });
-    if (USER?.id) {
-      logger.info(`User Obtained. ${USER?.id}`);
-    } else {
-      logger.info("No user found");
-    }
-    return {
-      ...initContextCache(),
-      DB,
-      USER,
-      RESEND,
-      PURCHASE_CALLBACK_URL,
-      GOOGLE_PHOTOS_IMPORT_QUEUE,
-      MAIL_QUEUE,
-      GET_SANITY_CLIENT,
-      GET_STRIPE_CLIENT,
-      GET_MERCADOPAGO_CLIENT,
-    } satisfies Context;
-  },
+  context: creageGraphqlContext,
 });
 
 export default {
   fetch: async (req: Request, env: Env, ctx: ExecutionContext) => {
-    attachPossibleUserIdFromJWT(req);
+    logPossibleUserIdFromJWT(req);
     logger.info("🏁 — Initialize Request");
     const response = await yoga.fetch(
       // @ts-expect-error Los tipos de yoga están mal
